# docs/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 星球出机 API
  description: |
    契约快照占位文件。后端工程接入 springdoc-openapi 后，用 /v3/api-docs.yaml 导出覆盖本文件。
    全局约定（鉴权、错误体、分页、金额单位）见同级目录 ../00-接口通则.md
  version: 0.0.0
servers:
  - url: https://api.example.com/api/v1
    description: 替换为实际 API 基址（含 /api/v1）
...
paths:
...
  /admin/community/posts:
    get:
      tags: [后台管理]
      summary: 后台社区帖子列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: status
          schema: { type: string, enum: [AUDITING, APPROVED, REJECTED, MANUAL_REVIEW, REVOKED] }
        - in: query
          name: circleId
          schema: { type: integer, format: int64 }
        - in: query
          name: keyword
          schema: { type: string }
      responses:
        '200':
          description: 帖子分页
  /admin/community/posts/{id}/revoke:
    post:
      tags: [后台管理]
      summary: 后台撤销帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 撤销成功
  /admin/community/posts/{id}/restore:
    post:
      tags: [后台管理]
      summary: 后台恢复帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 恢复成功
  /admin/community/posts/{id}/pin:
    post:
      tags: [后台管理]
      summary: 后台置顶帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 置顶成功
  /admin/community/posts/{id}/unpin:
    post:
      tags: [后台管理]
      summary: 后台取消置顶帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 取消成功
  /admin/community/topics:
    get:
      tags: [后台管理]
      summary: 后台社区话题列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: keyword
          schema: { type: string }
      responses:
        '200':
          description: 话题分页
  /admin/community/topics/{id}/revoke:
    post:
      tags: [后台管理]
      summary: 后台撤销话题下的可见帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 撤销成功
  /admin/community/topics/{id}/restore:
    post:
      tags: [后台管理]
      summary: 后台恢复话题下已撤销帖子
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 恢复成功
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 小程序登录后获得的 accessToken
...
security:
  - bearerAuth: []